requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.1
asynctest==0.13.0
atomicwrites==1.4.1
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.5
cryptography==37.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.15.1
Flask==2.0.3
flatbuffers==1.12
frozenlist==1.2.0
gast==0.3.3
google-auth==2.6.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.42.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.2.0
iniconfig==1.1.1
ipykernel==6.9.1
ipython==7.31.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyterlab-pygments==0.1.2
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.1.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.21.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.1.2
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.20
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pywin32==302
pywinpty==2.0.2
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.3
Send2Trash==1.8.0
setuptools==61.2.0
six==1.16.0
soupsieve==2.3.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.6.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
yarl==1.6.3
zipp==3.8.0
